+++ CHANGELOG.md
@@ -13,3 +13,5 @@ Here is a list of changes for each of the releases.
 - Formatter no longer forces BOM into files since `MASM` doesn't recognize it
 - Line breaks `LF` and `CRLF` are preserved instead of forcing `CRLF`
 - Added `--linebreaks` option to force specific line breaks
+- Added `--compact` option to optionally replace surplus blank lines with single blank line
+- Improved how procedures are separated and compacted